+++ Camera/include/camera_config.hh
@@ -11,16 +11,19 @@
 namespace holovibes::settings
 {
 // Very ugly to redefine the version in the "camera" folder but we have no choice if we want to define it only once.
-#define __HOLOVIBES_VERSION__ "13.1.2"
+#define __HOLOVIBES_VERSION__ "13.3.2"
 #define __APPNAME__ "Holovibes"
 
 #define __CAMERAS_CONFIG_FOLDER__ "cameras_config"
+#define __PRESET_FOLDER__ "preset"
 
 #define __APPDATA_HOLOVIBES_FOLDER__ (std::filesystem::path(getenv("APPDATA")) / __APPNAME__)
 #define __CONFIG_FOLDER__ (__APPDATA_HOLOVIBES_FOLDER__ / __HOLOVIBES_VERSION__)
 
 #define __CAMERAS_CONFIG_FOLDER_PATH__ (__CONFIG_FOLDER__ / __CAMERAS_CONFIG_FOLDER__)
+#define __PRESET_FOLDER_PATH__ (__CONFIG_FOLDER__ / __PRESET_FOLDER__)
 
 #define __CAMERAS_CONFIG_REFERENCE__ ((std::filesystem::path("AppData") / "cameras_config"))
+#define __PRESET_REFERENCE__ ((std::filesystem::path("AppData") / "preset"))
 
 } // namespace holovibes::settings