+++ Holovibes/includes/settings/settings.hh
@@ -163,9 +163,7 @@ DECLARE_SETTING(ContrastUpperThreshold, float)
 DECLARE_SETTING(RenormConstant, unsigned)
 DECLARE_SETTING(CutsContrastPOffset, size_t)
 DECLARE_SETTING(RecordOnGPU, bool)
-#ifndef _DEBUG
 DECLARE_SETTING(BenchmarkMode, bool)
-#endif
 // ex ComputeCache
 DECLARE_SETTING(BatchSize, uint)
 DECLARE_SETTING(TimeStride, uint)