+++ Holovibes/sources/gui/windows/lightui.cc
@@ -25,17 +25,16 @@ LightUI::LightUI(QWidget* parent, MainWindow* main_window, ExportPanel* export_p
     , z_distance_subscriber_("z_distance", std::bind(&LightUI::actualise_z_distance, this, std::placeholders::_1))
     , record_start_subscriber_("record_start", std::bind(&LightUI::on_record_start, this, std::placeholders::_1))
     , record_end_subscriber_("record_stop", std::bind(&LightUI::on_record_stop, this, std::placeholders::_1))
+    , record_finished_subscriber_("record_finished", [this](bool success)
+                                       {
+                                           reset_start_button();
+                                           reset_record_progress_bar();
+                                       })
 {
     ui_->setupUi(this);
 
-    connect(ui_->OutputFileBrowseToolButton, &QPushButton::clicked, this, &LightUI::browse_record_output_file_ui);
-    connect(ui_->OutputFileNameLineEdit, &QLineEdit::textChanged, this, &LightUI::set_record_file_name);
-    connect(ui_->startButton, &QPushButton::toggled, this, &LightUI::start_stop_recording);
-    connect(ui_->actionConfiguration_UI, &QAction::triggered, this, &LightUI::open_configuration_ui);
-    connect(ui_->ZSpinBox, &QSpinBox::valueChanged, this, &LightUI::z_value_changed_spinBox);
-    connect(ui_->ZSlider, &QSlider::valueChanged, this, &LightUI::z_value_changed_slider);
-
     ui_->startButton->setStyleSheet("background-color: rgb(50, 50, 50);");
+    // ui_->startButton->setShortcut(Qt::CTRL + Qt::Key_R);  #FIXME: This shortcut is not working, even though it works for MainWindow
 }
 
 LightUI::~LightUI()
@@ -51,12 +50,11 @@ void LightUI::showEvent(QShowEvent* event)
     visible_ = true;
 }
 
-void LightUI::actualise_record_output_file_ui(const QString& file_path)
+void LightUI::actualise_record_output_file_ui(const std::filesystem::path file_path)
 {
-    ui_->OutputFilePathLineEdit->setText(QFileInfo(file_path).path());
-    auto file_name = QFileInfo(file_path).fileName();
+    ui_->OutputFilePathLineEdit->setText(QString::fromStdString(file_path.parent_path().string()));
     // remove the extension from the filename
-    ui_->OutputFileNameLineEdit->setText(file_name.left(file_name.lastIndexOf('.')));
+    ui_->OutputFileNameLineEdit->setText(QString::fromStdString(file_path.stem().string()));
 }
 
 void LightUI::actualise_z_distance(const double z_distance)
@@ -67,12 +65,10 @@ void LightUI::actualise_z_distance(const double z_distance)
     ui_->ZSlider->setValue(static_cast<int>(std::round(z_distance * 1000)));
 }
 
-void LightUI::z_value_changed_spinBox(int z_distance)
+void LightUI::z_value_changed(int z_distance)
 {
-    api::set_z_distance(static_cast<double>(z_distance) / 1000.0f);
+    api::set_z_distance(static_cast<float>(z_distance) / 1000.0f);
 }
-
-void LightUI::z_value_changed_slider(int z_distance) { api::set_z_distance(static_cast<double>(z_distance) / 1000.0f); }
 
 void LightUI::browse_record_output_file_ui()
 {
@@ -80,15 +76,17 @@ void LightUI::browse_record_output_file_ui()
     //! not know about the MainWindow and the ExportPanel. It should only know about the API. One way to fix it is to
     //! create a new browser in this class and then use notify to send the file path to the API (and synchronize the API
     //! with the file path).
-    auto file_path = export_panel_->browse_record_output_file();
-    ui_->OutputFilePathLineEdit->setText(QFileInfo(file_path).path());
-    auto file_name = QFileInfo(file_path).fileName();
-    // remove the extension from the filename
-    ui_->OutputFileNameLineEdit->setText(file_name.left(file_name.lastIndexOf('.')));
-}
+    std::filesystem::path file_path(export_panel_->browse_record_output_file().toStdString());
+    std::string file_path_str = file_path.string();
+    std::replace(file_path_str.begin(), file_path_str.end(), '/', '\\');
+    ui_->OutputFilePathLineEdit->setText(QString::fromStdString(file_path_str));
 
-void LightUI::set_record_file_name(const QString& filename)
+    // remove the extension from the filename
+    ui_->OutputFileNameLineEdit->setText(QString::fromStdString(file_path.stem().string()));
+}
+void LightUI::set_record_file_name()
 {
+    QString filename = ui_->OutputFileNameLineEdit->text();
     // concatenate the path with the filename
     std::filesystem::path path(ui_->OutputFilePathLineEdit->text().toStdString());
     std::filesystem::path file(filename.toStdString());
@@ -134,15 +132,9 @@ void LightUI::actualise_record_progress(const int value, const int max)
     ui_->recordProgressBar->setValue(value);
 }
 
-void LightUI::set_visible_record_progress(bool visible)
+void LightUI::reset_record_progress_bar()
 {
-    if (visible)
-        ui_->recordProgressBar->show();
-    else
-    {
-        ui_->recordProgressBar->reset();
-        ui_->recordProgressBar->setFormat("Idle");
-    }
+    set_recordProgressBar_color(QColor(10, 10, 10), "Idle");
 }
 
 void LightUI::set_recordProgressBar_color(const QColor& color, const QString& text)
@@ -177,6 +169,13 @@ void LightUI::open_configuration_ui()
     main_window_->show();
     this->hide();
     visible_ = false;
+}
+
+void LightUI::set_preset()
+{
+    std::filesystem::path dest = __PRESET_FOLDER_PATH__ / "preset.json";
+    main_window_->reload_ini(dest.string());
+    LOG_INFO("Preset loaded");
 }
 
 void LightUI::closeEvent(QCloseEvent* event) { main_window_->close(); }